initial add of aframe elements
Currently all of the geo and maps are versioned in GIT which could end up being bloated depending on how many versions we go through - which will be many. I'm currently rewrapping the built in shader to get lightmap support. The shader rewrite (which is not in thus commit) is a texture memory hog right now, but will allow for us to separate out baked lighting from textures. In order to fully separate the two we'll need to add secondary UVs to the geometry. In order for it to be more memory efficient it needs to be able to share texture maps between material instances. 
Other work TBD:
 - constraining navigation to a mesh - most AFrame navigation is currently either unconstrained or based on physics collisions and gravity.
 - pulling the geo and image assets from urls so that they don't take up GIT space.
 - current lighting is temp and only direct and indirect diffuse, needs aframe live lights and reflection maps. 
 - model needs side rooms, railings, control room, light r